a bit for refactoring
